pyproject.toml

[project]
name = "myra"
version = "0.1.1"
description = "A library to count the colors in an image."
authors = [
    {name = "daniel daum", email = "daniel_daum@outlook.com"},
]
dependencies = [
    "pillow>=9.5.0",
    "polars>=0.17.13",
    "isort>=5.12.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.0.0",
    "pytest-sugar>=0.9.7",
    "black>=23.3.0",
    "isort>=5.12.0",
    "ruff>=0.0.267",
    "pandas>=2.0.1",
]
